requirements.txt
asgiref==3.2.10
bcrypt==3.2.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.1.1
Django==3.1.2
django-crispy-forms==1.9.2
django-rest-knox==4.1.0
djangorestframework==3.12.1
httpie==2.2.0
idna==2.10
pkg-resources==0.0.0
pycparser==2.20
Pygments==2.7.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.25.10

